+++ src/pages/trade/TradeOrderPage.vue
@@ -7,8 +7,8 @@
     </div>
 
     <div class="flex-1 overflow-y-auto">
-      <HogaPageComponent v-if="currentFundingStatus === 'askingPrice'" />
-      <StockChartComponent v-if="currentFundingStatus === 'stockChart'" />
+      <OrderbookChart v-if="currentFundingStatus === 'askingPrice'" />
+      <TradeHistoryChart v-if="currentFundingStatus === 'stockChart'" />
     </div>
     <BuyAndSellAccodian class="fixed bottom-16 left-0 right-0 max-w-md mx-auto">
     </BuyAndSellAccodian>
@@ -18,12 +18,12 @@
 <script setup>
 import { ref } from 'vue'
 import BaseTab from '@/components/common/Tab/BaseTab.vue'
-import StockChartComponent from '@/components/trade/StockChartComponent.vue'
 import BlankLayout from '@/layouts/BlankLayout.vue'
 import DetailHeader from '@/layouts/DetailHeader.vue'
 import BuyAndSellAccodian from '@/components/trade/BuyAndSell/BuyAndSellAccodian.vue'
-import HogaPageComponent from '@/components/trade/Hoga/HogaPageComponent.vue'
 import CurrentPrice from '@/components/trade/CurrentPrice.vue'
+import TradeHistoryChart from '@/components/trade/TradeHistoryChart.vue'
+import OrderbookChart from '@/components/trade/Hoga/OrderbookChart.vue'
 
 const fundingStatusTabs = [
   { label: '호가', value: 'askingPrice' },
